Description_national_city_bank
The National City Bank research and modeling project was to create a customer propensity model for a new product - a line of credit against a household's used car. Using historical data from 4000 previous calls and mailings, I performed EDA, create a propensity model, and evaluate it to identify the next 100 customers from a prospective customer list to contact. Bankers will call and direct mail will be sent to households with the highest probability of accepting the offer.

I also analyzed the customer profile for both historical and top prospective customers, and use variable importance and sound EDA to aid the presentation.