First project using R, description
For this project, we acted as a member of the business intelligence team at OkCupid.com, and have been tasked with conducting an in-depth exploratory analysis of site users to create micro-segments and personas for future marketing campaigns. However, it is important to note that interesting data correlations may not always be beneficial in a marketing context. The goal is to identify significant insights that can lead to actionable marketing strategies.

This assignment was to clean the data, enrich it with supplementary data, and identify at least four interesting insights from the user data. All relevant cleaning, enriching, and EDA steps, along with the four insightful data nuances.